pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "zuviscribe"
version = "0.1.0"
description = "Turn any non-animated YouTube video into a creative animated rendition."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Manish Suvarna"},
]
dependencies = [
    "yt-dlp>=2024.1",
    "openai-whisper>=20231117",
    "ollama>=0.1.6",
    "langchain>=0.1",
    "langchain-community>=0.0.10",
    "Pillow>=10.0",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "ffmpeg-python>=0.2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "click>=8.1",
    "tqdm>=4.66",
    "TTS>=0.22",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.1",
    "fastapi>=0.108",
    "uvicorn>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
zuviscribe = "zuviscribe.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["zuviscribe*", "agents*", "tools*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
